+++ backend/daoumarket/src/main/java/com/daoumarket/dto/User.java
@@ -10,8 +10,8 @@ import lombok.ToString;
 @Setter
 @ToString
 public class User {
-	private long userId;
-	private long empNum;
+	private int userId;
+	private int empNum;
 	private String name;
 	private String password;
 	private String department;
