+++ backend/daoumarket/src/main/java/com/daoumarket/service/IFavoriteService.java
@@ -8,5 +8,5 @@ import com.daoumarket.dto.Favorite;
 public interface IFavoriteService {
 	public ResponseEntity<BasicResponse> insertFavorite(Favorite favorite);
 	public ResponseEntity<BasicResponse> deleteFavorite(long favoriteId);
-	public ResponseEntity<BasicResponse> getFavoriteList(long userId);
+	public ResponseEntity<BasicResponse> getFavoriteList(int userId);
 }
