+++ backend/daoumarket/src/main/java/com/daoumarket/service/IUserService.java
@@ -8,7 +8,7 @@ import com.daoumarket.dto.UserLoginRequest;
 
 public interface IUserService {
 	public ResponseEntity<BasicResponse> insertUser(User user);
-	public ResponseEntity<BasicResponse> getEmpNum(long empNum);
+	public ResponseEntity<BasicResponse> getEmpNum(int empNum);
 	public ResponseEntity<BasicResponse> getUserLogin(UserLoginRequest userLogin);
 	public ResponseEntity<BasicResponse> updateUser(User user);
 }
